+++ website_dating/static/src/js/notification.js
@@ -1,0 +1,11 @@
+odoo.define('website_dating.dating_notification', function (require) {
+"use strict";
+
+$(document).on('click', '#dating_notif_link', function(){
+
+    $("#dating_notifications").css("display","block");
+
+});
+
+
+});