+++ website_dating_events/models/res_partner_dating_like.py
@@ -4,8 +4,8 @@ _logger = logging.getLogger(__name__)
 
 from openerp import api, fields, models
 
-class ResPartnerLikeEvent(models.Model):
+class ResPartnerDatingLikeEvent(models.Model):
 
-    _inherit = "res.partner.like"
+    _inherit = "res.partner.dating.like"
 
     event_id = fields.Many2one('event.event', string="Event")