+++ website_dating/doc/changelog.rst
@@ -1,3 +1,7 @@
+v0.3.1
+======
+* Notification system
+
 v0.3
 ====
 * Message and like system