+++ website_dating_events/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.1
+======
+* Change to use notification system
+
 v1.0
 ====
 * Initial Release